requirements.txt
streamlit==1.41.1
snowflake-connector-python==3.14.0
requests==2.32.3
pandas==2.2.3
